refactor(user invites): refactored user invite.
